build: Add ignore config for spotbugs in jenkins
Fixes #696

Jenkins build is failing due to incompatible jenkins/spotbugs versions.
This has to be fixed in the jenkins server by a jenkins upgrade. Util
that is completed this temporary fix is added to support a jenkins pre
build script which will remove spotbugs plugin definition from pom file
before the build.