project: Add .keep to interceptors dir
Fixes #715

If a project is checkedin into a VCS wihtout any interceptors, next
dev who is going to checkout the project and start working will get
into an error due to unavailable interceptors directory. This
happens because empty directories are not checked into VCS. To avoid
this we added a .keep file to interceptors directory.