cli: build: Use all project extensions
Fixes #764

If any custom extensions are provided inside project's extensions
directory those will be copied to target directory as sources for
building the project. This make it possible to add additional
filters to a specific project. However to make this useful, user
will have to update the listneres.bal in TOOLKIT resources.