bin: Use internal BALLERINA_HOME
Fixes #810

If $BALLERINA_HOME/bin is appended to PATH variable env will use the
first exec found on the PATH variable. If $BALLERINA_HOME is already
defined by the user, then that version of ballerina will be used
ignoring the internal ballerina installation of mgw.
Therefore we should always prepend the $BALLERINA_HOME/bin value to
PATH variable so that internal version is picked up at the evaluation
of PATH variable.