etcd: Fix issue in endpoint reinit
Fixes: 923

@index is used to define the endpoint index. However for http endpoints
(when the endpoint is not load balancer or failover etc.), @index
doesn't mean anything as this code line is not inside a loop. In this
case Mustache was resolving the value of @index to 0 and sometimes 1.
Which cause the etcd implementation to fail when trying to find the
configured endpoint for some resources.