docker: Update base image
1. Ballerina jdk requires a glib based base image. However default
alpine image is built with musl not glibc. This causes mgw startup
issues when http2 features are enabled. To resolve this we used a
glib based alpine image 'alpine-glibc'. This image is shared in
https://github.com/Docker-Hub-frolvlad/docker-alpine-glibc

2. COPY command creates an unwanted layer in the docker image.
Therefore we now use wget to pull the realeased runtime dist from
github. This will increase both efficiency and portability of the
Dockerfile