runtime: cli: Fix error log in startup
Fixes #1107

No need to pass the command name to tools lib now. It only contain
one command at the moment.
CONF_OUT_FILE is not written to disk if observability configuration
is not found on the micro-gw.conf file. Therefore error log should
not be printed if CONF_OUT_FILE is not found