cli: build: Fix build status message
Having the build status message in a separate line make thing look
unproffessional. Therefore this commit take required actions to
get the status message in to the same line as " Generating sources"
message.
Proto compiler download messages were replaced with progress dots,
if we need to debug any issue with the download we can request for
the log while as it contains the file name and few other info.