core: Pass interceptor response to client
When interceptor responds with 401, 403 like error responses,
gateway should avoid sending it's own response to the client.
This fix is an improvement to the fix done in 35a07d7 where
that fix fails to identify responses coming from an interceptor.
When an interceptor responds to the client, the developer should
set RESPOND_DONE property to the invocationContext before calling
respond(). Otherwise mgw will not be able to recognize the
response as a response coming from an interceptor.