tool: Read config from args and env
Fixes #1201

Startup tool library was only reading the jmx_port config from the
config file. When config is provided as program arg or env varriable
it was not picked up.
This fix will anable looking up for both program args and env vars
for the configuration. This is implemented in the tool lib to make
it general for both *nix and wnidows platforms