readme: Update and simplify the contents
There were many contents in the readme that duplicated the contents
of the official docs. This makes things hard to maintain and at
times devs forget to update the readme contents for each release.
To avoid the duplicated documentation, readme is simplified to
point at the official documentation where it is required.