throttle: Sync keytemplate msgs
When keytemplate is updated (redeployed) in APIM order of the
jms messages received by mgw is not guaranteed. If first msg
is a 'add' and the next one is a 'remove', updated keyemplate
is lost in mgw side. To resolve this we sync the msg order by
looking at the jms msg timestamp.