Arrays of type string, uint, int, float can be resolved from env variables.
[category]
stringArray = ["foo"]

configuration can be updated with

export cc_category_stringArray="foo, bar"

then the assigned values for the array would be foo and bar.

For integers, and floats also just provide the values as comma separated values. If one of the elements are
not parsed properly, the array will not be updated from the environment at all.