Provide capability to have apis deployed during the adapter start
Via this feature, the users can mount their API projects (which were created by apictl init method or
apictl export api command) can be mounted to the adapter container (default path is /home/wso2/artifacts/apis).
Those APIs will be availble during the startup and readiness probe would be set only after applying all the APIs.

API artifacts can remain in either as a directory or in zipped format. And do not include them in nested structures.
If a certain API artifact is faulty, then an error message will be printed and the adapter will start with remaining
set of APIs.

This feature will only work if the control plane is disabled.

If a different file path needs to be used within the container, update the config.toml with

[adapter]
artifactsDirectory = "<custom-path>"