Custom configurations for Filters
If the custom filter needs to be configured prior to deployment, the config properties map can added to the filters.

[[enforcer.filters]]
    # ClassName of the filter
    className = "org.wso2.choreo.connect.custom.CCCustomFilter"
    # Position of the filter within final filter-chain
    position = 3
    # Custom Configurations
    [enforcer.filters.configProperties]
        testConfig1 = "testValue1"

Inside the custom filter init method, the properties can be set to filter object. The filter object is created for each API Instance.
So once it is set in the filter instance during init method, it can be utilized at handlerequest method as user wish.