adp: trace: Add trace cluster
This will add support to add globally available cluster's to
envoy clusters array. Currently we have support to add dynamic
clusters of API endpoints but there are some clusters that
should be added dynamically based on configs, also if enabled
by config, those clusters should be available globally on all
labels.
To implement this we are keeping a global array of clusters &
endpoints similar to other envoy resources such as routes.