adp: defversion: Support default version with CP
default version change event is coming as an API_UPDATE type event from
CP. This event doesn't contain ebough information for CC to figureout if
it should redeploy the API. Hence the `Action` property is introduced
to API_UPDATE event.
API_UPDATE event doesn't contain the gatewayLabels property, therefore
we are using intenal maps to find the deployed envs of an API when
default version change event is received.