adp: Fix integration test failure
panic error was returned randomly when APIListMap was lookedup
to get `isDefaultVersion` value. This can happen when this lookup
happens during normal API deployment events. Reason is this path
doesn't (and its not required to) synchronously update APIMeata
before fetching runtime artifact.
Since runtime artifact is anyway updated with `isDefaultVersion`
when API deployement happens, It's not required to have APIMeta
 up to date before fetching runtime artifact.