Parse the JWT payload with a JSON parser during JWT based throttle validations.
When throttling based on JWT claims, we parsed the payload of the JWT similar to a SWT until now. This was a direct transition from the json to a Map<String, String>. This did not require initializing a JSON parser and upto now we only had come accross use cases where the claim value was a string, and not an object. Recently, an update has been released for API-M 4.1.0 where the access token generated from developer portal includes the claim "realm" with a value that is an object (Ex. {\n "signing_tenant": "carbon.super" \n}). Hence, the payload must be considered as an actual JWT token which may include an JSON object as the claim value.