Include the original vhost in access logs
Original vhost (authority header received from request) has to be included in the access log to filter out the gateway environment.

Query params will be available in the original request path as well as in the backend request path in access logs. User is responsible for not sending sensitive data with query params.

Signed-off-by: Renuka Fernando <renukapiyumal@gmail.com>